Network Up/Down & Cleanup
Two features are now working: network upload speed and network download speed. In result_pretty, the units for the speed dynamically change in order to make it fit a specific size, which is pretty cool. Futhermore, the Delta unit type now uses VecDeque, which makes code a bit more elegant.